+++ phantasm/crawler.go
@@ -1,4 +1,4 @@
-package steaminvcrawler
+package phantasm
 
 import (
 	"bytes"
