+++ phantasm/example/main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"fmt"
 
-	"github.com/kudarap/dotagiftx/tools/steaminvcrawler"
+	"github.com/kudarap/dotagiftx/phantasm"
 )
 
 func main() {
 	args := map[string]interface{}{
 		"steam_id": "76561198088587178",
 	}
-	v := steaminvcrawler.Main(args)
+	v := phantasm.Main(args)
 	fmt.Println(v)
 }
