+++ phantasm/testdata/inventory_paginated_1.json
