+++ phantasm/testdata/inventory_paginated_2.json
