+++ phantasm/testdata/inventory_paginated_3.json
