+++ phantasm/testdata/inventory_paginated_4.json
